separate test cases into separate files
